feat: 댓글 좋아요 비회원 예외 처리

+++ src/main/java/com/spartanullnull/otil/domain/like/controller/CommentLikeController.java
@@ -3,6 +3,7 @@ package com.spartanullnull.otil.domain.like.controller;
 import com.spartanullnull.otil.domain.like.service.CommentLikeService;
 import com.spartanullnull.otil.security.Impl.UserDetailsImpl;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.DeleteMapping;
@@ -11,6 +12,7 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.server.ResponseStatusException;
 
 @RestController
 @RequestMapping("/api/posts")
@@ -25,6 +27,12 @@ public class CommentLikeController {
         @PathVariable Long postId,
         @PathVariable Long commentId,
         @AuthenticationPrincipal UserDetailsImpl userDetails) {
+
+        // 로그인한 사용자 체크
+        if (userDetails == null) {
+            throw new ResponseStatusException(HttpStatus.UNAUTHORIZED, "로그인한 사용자만 댓글 좋아요를 누를 수 있습니다.");
+        }
+
         commentLikeService.createLikeForComment(userDetails.getUser().getId(), commentId);
         return ResponseEntity.ok("댓글 좋아요 성공!");
     }
@@ -35,6 +43,12 @@ public class CommentLikeController {
         @PathVariable Long postId,
         @PathVariable Long commentId,
         @AuthenticationPrincipal UserDetailsImpl userDetails) {
+
+        // 로그인한 사용자 체크
+        if (userDetails == null) {
+            throw new ResponseStatusException(HttpStatus.UNAUTHORIZED, "로그인한 사용자만 댓글 좋아요를 취소할 수 있습니다.");
+        }
+
         commentLikeService.deleteLikeForComment(userDetails.getUser().getId(), commentId);
         return ResponseEntity.ok("댓글 좋아요 취소!");
     }
